+++ dRowAudio/maths/dRowAudio_BezierCurve.h
@@ -40,7 +40,7 @@ namespace BezierCurve
 	 */
 	inline float slopeFromT (float t, float A, float B, float C)
     {
-		float dtdx = 1.0 / (3.0*A*t*t + 2.0*B*t + C); 
+		float dtdx = 1.0f / (3.0f*A*t*t + 2.0f*B*t + C); 
 		return dtdx;
 	}
 	
@@ -91,16 +91,16 @@ namespace BezierCurve
 		// by Don Lancaster, SYNERGETICS Inc. 
 		// http://www.tinaja.com/text/bezmath.html
 		
-		float epsilon = 0.00001;
-		a = jmax(0.0f, jmin(1.0f, a)); 
-		b = jmax(0.0f, jmin(1.0f, b)); 
+		float epsilon = 0.00001f;
+		a = jmax (0.0f, jmin (1.0f, a)); 
+		b = jmax (0.0f, jmin (1.0f, b)); 
 		if (a == 0.5){
 			a += epsilon;
 		}
 		
 		// solve t from x (an inverse operation)
 		float om2a = 1 - 2*a;
-		float t = (sqrt(a*a + om2a*x) - a)/om2a;
+		float t = ((float) sqrt (a*a + om2a*x) - a)/om2a;
 		float y = (1-2*b)*(t*t) + (2*b)*t;
 		return y;
 	}
@@ -113,14 +113,14 @@ namespace BezierCurve
 	static float cubicBezier (float x, float a, float b, float c, float d)
 	{
 		
-		float y0a = 0.00; // initial y
-		float x0a = 0.00; // initial x 
-		float y1a = b;    // 1st influence y   
-		float x1a = a;    // 1st influence x 
-		float y2a = d;    // 2nd influence y
-		float x2a = c;    // 2nd influence x
-		float y3a = 1.00; // final y 
-		float x3a = 1.00; // final x 
+		float y0a = 0.00f; // initial y
+		float x0a = 0.00f; // initial x 
+		float y1a = b;     // 1st influence y   
+		float x1a = a;     // 1st influence x 
+		float y2a = d;     // 2nd influence y
+		float x2a = c;     // 2nd influence x
+		float y3a = 1.00f; // final y 
+		float x3a = 1.00f; // final x 
 		
 		float A =   x3a - 3*x2a + 3*x1a - x0a;
 		float B = 3*x2a - 6*x1a + 3*x0a;
@@ -158,11 +158,11 @@ namespace BezierCurve
 	{
 		
 		float y = 0;
-		float epsilon = 0.00001;
-		float min_param_a = 0.0 + epsilon;
-		float max_param_a = 1.0 - epsilon;
-		float min_param_b = 0.0 + epsilon;
-		float max_param_b = 1.0 - epsilon;
+		float epsilon = 0.00001f;
+		float min_param_a = 0.0f + epsilon;
+		float max_param_a = 1.0f - epsilon;
+		float min_param_b = 0.0f + epsilon;
+		float max_param_b = 1.0f - epsilon;
 		a = jmax (min_param_a, jmin (max_param_a, a));
 		b = jmax (min_param_b, jmin (max_param_b, b));
 		
@@ -178,8 +178,8 @@ namespace BezierCurve
 		
 		// arbitrary but reasonable 
 		// t-values for interior control points
-		float t1 = 0.3;
-		float t2 = 0.7;
+		float t1 = 0.3f;
+		float t2 = 0.7f;
 		
 		float B0t1 = B0(t1);
 		float B1t1 = B1(t1);
